+++ src/core/array-list.ts
