+++ src/core/primitive-allocator.ts
